Move common on next unique id fetch actions to one class
