+++ rdt-app/src/main/java/io/ona/rdt_app/interactor/PatientRegisterFragmentInteractor.java
@@ -1,13 +1,11 @@
 package io.ona.rdt_app.interactor;
 
-import android.app.Activity;
 import android.os.AsyncTask;
 import android.util.Log;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-import org.apache.commons.lang3.StringUtils;
 import org.joda.time.DateTime;
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -15,7 +13,6 @@ import org.json.JSONObject;
 import org.smartregister.clientandeventmodel.Client;
 import org.smartregister.clientandeventmodel.Event;
 import org.smartregister.domain.LocationProperty;
-import org.smartregister.domain.UniqueId;
 import org.smartregister.domain.db.EventClient;
 import org.smartregister.domain.tag.FormTag;
 import org.smartregister.exception.JsonFormMissingStepCountException;
@@ -28,21 +25,15 @@ import org.smartregister.util.PropertiesConverter;
 import java.util.Calendar;
 import java.util.Collections;
 
-import io.ona.rdt_app.R;
 import io.ona.rdt_app.application.RDTApplication;
 import io.ona.rdt_app.callback.OnFormSavedCallback;
-import io.ona.rdt_app.callback.OnUniqueIdFetchedCallback;
-import io.ona.rdt_app.model.Patient;
-import io.ona.rdt_app.util.FormLaunchArgs;
+import io.ona.rdt_app.util.FormLauncher;
 import io.ona.rdt_app.util.RDTJsonFormUtils;
-import timber.log.Timber;
 
-import static com.vijay.jsonwizard.utils.Utils.showToast;
 import static io.ona.rdt_app.util.Constants.DETAILS;
 import static io.ona.rdt_app.util.Constants.DOB;
 import static io.ona.rdt_app.util.Constants.ENCOUNTER_TYPE;
 import static io.ona.rdt_app.util.Constants.ENTITY_ID;
-import static io.ona.rdt_app.util.Constants.Form.RDT_TEST_FORM;
 import static io.ona.rdt_app.util.Constants.METADATA;
 import static io.ona.rdt_app.util.Constants.PATIENTS;
 import static io.ona.rdt_app.util.Constants.PATIENT_AGE;
@@ -57,7 +48,7 @@ import static org.smartregister.util.JsonFormUtils.getString;
 /**
  * Created by Vincent Karuri on 13/06/2019
  */
-public class PatientRegisterFragmentInteractor implements OnUniqueIdFetchedCallback {
+public class PatientRegisterFragmentInteractor extends FormLauncher {
 
     private static final Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
             .registerTypeAdapter(DateTime.class, new DateTimeTypeConverter())
@@ -147,32 +138,5 @@ public class PatientRegisterFragmentInteractor implements OnUniqueIdFetchedCallb
         org.smartregister.domain.db.Event dbEvent = gson.fromJson(eventJson.toString(), org.smartregister.domain.db.Event.class);
 
         return new EventClient(dbEvent, dbClient);
-    }
-
-    public void launchForm(Activity activity, String formName, Patient patient) throws JSONException {
-        if (patient != null) {
-            FormLaunchArgs args = new FormLaunchArgs();
-            args.withActivity(activity)
-                    .withFormJsonObj(new JSONObject(formName))
-                    .withPatient(patient);
-            formUtils.getNextUniqueId(args, this);
-        } else {
-            formUtils.launchForm(activity, formName, patient, null);
-        }
-    }
-
-    @Override
-    public synchronized void onUniqueIdFetched(FormLaunchArgs args, UniqueId uniqueId) {
-        try {
-            String rdtId = uniqueId.getOpenmrsId();
-            Activity activity = args.getActivity();
-            if (!StringUtils.isBlank(rdtId)) {
-                formUtils.launchForm(activity, RDT_TEST_FORM, args.getPatient(), rdtId);
-            } else {
-                showToast(activity, activity.getString(R.string.unique_id_fetch_error_msg));
-            }
-        } catch (JSONException e) {
-            Timber.e(TAG, e.getStackTrace().toString());
-        }
     }
 }
